Move vg stats to a subcommand
